split risk-advisor and proxy into separate binaries
