+++ build.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+go install cmd/proxy
+go install cmd/riskadvisor
+
+# now you can run proxy using proxy command and riskadvisor using riskadvisor command
