+++ cmd/proxy/app/riskadvisorHandler/server.go
@@ -10,7 +10,7 @@ import (
 
 	"k8s.io/kubernetes/pkg/api"
 
-	"github.com/Prytu/risk-advisor/podprovider"
+	"github.com/Prytu/risk-advisor/cmd/proxy/app/podprovider"
 )
 
 type RiskAdvisorHandler struct {
