+++ cmd/proxy/main.go
@@ -1,17 +1,18 @@
 package main
 
 import (
-	"github.com/Prytu/risk-advisor/podprovider"
-	"github.com/Prytu/risk-advisor/proxy"
-	"github.com/Prytu/risk-advisor/proxy/riskadvisorHandler"
-	"github.com/Prytu/risk-advisor/riskadvisor"
-	"k8s.io/kubernetes/pkg/api"
+	"log"
 	"net/http"
+
+	"k8s.io/kubernetes/pkg/api"
+
+	"github.com/Prytu/risk-advisor/cmd/proxy/app"
+	"github.com/Prytu/risk-advisor/cmd/proxy/app/podprovider"
+	"github.com/Prytu/risk-advisor/cmd/proxy/app/riskadvisorHandler"
 )
 
 // read from somewhere
 const realApiserverURL = "http://localhost:8080"
-const riskAdvisorPort = ":9997"
 const proxyRACommunicationPort = ":9998"
 const proxySchedulerCommunicationPort = ":9999"
 
@@ -22,15 +23,15 @@ func main() {
 
 	raHandler := riskadvisorhandler.New(responseChannel, errorChannel, podProvider)
 
-	proxy, err := proxy.New(realApiserverURL, podProvider, responseChannel, errorChannel)
+	proxy, err := app.New(realApiserverURL, podProvider, responseChannel, errorChannel)
 	if err != nil {
 		panic(err)
 	}
 
-	// TODO: Ugly, but will be fixed in issue #11
-	riskAdvisor := riskadvisor.New("http://localhost" + proxyRACommunicationPort + "/advise")
+	log.Printf("Staring proxy with:\n\t- real apiserver URL: %v\n\t- scheduler communication port: %v"+
+		"\n\t- risk-advisor communication port: %v\n", realApiserverURL, proxySchedulerCommunicationPort,
+		proxyRACommunicationPort)
 
 	go http.ListenAndServe(proxyRACommunicationPort, raHandler)
-	go http.ListenAndServe(riskAdvisorPort, riskAdvisor)
 	http.ListenAndServe(proxySchedulerCommunicationPort, proxy)
 }
