+++ cmd/riskadvisor/app/model.go
@@ -1,10 +1,6 @@
-package riskadvisor
+package app
 
 import "k8s.io/kubernetes/pkg/api"
-
-type AdviceRequest struct {
-	Pod *api.Pod `json:"pod"`
-}
 
 type AdviceResponse struct {
 	Status string      `json:"status"`
