+++ pkg/model/model.go
@@ -1,0 +1,7 @@
+package model
+
+import "k8s.io/kubernetes/pkg/api"
+
+type AdviceRequest struct {
+	Pod *api.Pod `json:"pod"`
+}
